moved implementation details into dedicated subpackage
