+++ scripts/script_utils/deps.py
@@ -13,6 +13,7 @@
 # limitations under the License.
 #
 """Contains utils for working with dependencies, lock files, etc."""
+
 from copy import deepcopy
 from pathlib import Path
 from typing import Any
@@ -54,11 +55,11 @@ def remove_self_dependencies(pyproject: dict) -> dict:
 
     if "optional-dependencies" in project_metadata:
         for group in project_metadata["optional-dependencies"]:
-            project_metadata["optional-dependencies"][
-                group
-            ] = exclude_from_dependency_list(
-                package_name=package_name,
-                dependencies=project_metadata["optional-dependencies"][group],
+            project_metadata["optional-dependencies"][group] = (
+                exclude_from_dependency_list(
+                    package_name=package_name,
+                    dependencies=project_metadata["optional-dependencies"][group],
+                )
             )
 
     return modified_pyproject
