+++ src/arctictypes/_internal/freeze.py
@@ -17,11 +17,11 @@
 from collections.abc import Sequence
 from typing import Optional
 
-from arctictypes._converters import (
+from arctictypes._internal._converters import (
     STANDARD_CONVERTERS,
     Converter,
 )
-from arctictypes._resolve import (
+from arctictypes._internal.resolve import (
     ConverterNotFoundError,  # noqa: F401 - a shortcut
     get_converter_by_type,
 )
