+++ tests/test_freeze.py
@@ -15,8 +15,8 @@
 """Test the freeze function."""
 
 import pytest
-
 from arctictypes import freeze
+
 from tests.cases import INVALID_CASES, VALID_CASES, InvalidTestCase, ValidTestCase
